refactor(persona): split admission.rs (1225 LOC) into mod + recipes (#1208)
`persona/admission.rs` was 1225 LOC mixing the structural admission
gate, the IsMemorable trait, the v1 HeuristicIsMemorable recipe + its
policy tests, helpers, and the gate test suite. Split:

  admission/mod.rs     (985 LOC) — AdmissionGate::admit machinery,
                                    Candidate/Context/Config types,
                                    IsMemorable trait, envelope
                                    verification, seam recording, and
                                    the structural-gate test suite
                                    (replay, trust threshold, recipe
                                    error path, quarantine propagation,
                                    seam-emission invariants)
  admission/recipes.rs (326 LOC) — HeuristicIsMemorable struct + impl
                                    + 4 heuristic-policy tests
                                    (short_content, noise_phrase,
                                    duplicate, admit_synthesizes_engram)

`HeuristicIsMemorable` re-exported at the parent path via
`pub use recipes::HeuristicIsMemorable` — external callers see no API
change. Engram types previously imported privately from `super::engram`
are now re-exported `pub use` so submodules can reach them via `super::`.

Tests:
- `cargo check --features metal,accelerate -p continuum-core` clean.
- `cargo test --features metal,accelerate -p continuum-core --lib persona::admission`
  → 37 passed, 0 failed.

Closes #1208 — final slice. evaluator.rs done in #1242,
model_resolver.rs done in #1249, admission.rs done now.

Worktree-discipline note: this PR is the first work this session
authored from a proper `airc lane create` worktree rather than the
shared root checkout, after Joel called out that branch swaps in the
shared root were stomping uncommitted work.
